refactor(networks,tests): Remove ALL abstractions
- Moved everything into a single __init__.py file
- Removed forge.py as a side product of premature optimization
- Moved all tests back to tests/ directory
