+++ tests/test_cnn_build_model.py
