+++ tests/test_is_dict.py
@@ -1,4 +1,4 @@
-from template_nn.args_val import is_dict
+from template_nn import is_dict
 
 
 def test_is_dict_returns_expected_values():
